Books

+++ old-proje/src/main/java/az/devolopia/SpringJava20_Mubariz_Bayramov/service/BookService.java
@@ -1,5 +1,6 @@
 package az.devolopia.SpringJava20_Mubariz_Bayramov.service;
-import java.awt.print.Book;
+
+
 import java.io.BufferedReader;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
@@ -12,17 +13,19 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import az.devolopia.SpringJava20_Mubariz_Bayramov.config.MyConfig;
 import az.devolopia.SpringJava20_Mubariz_Bayramov.entity.BookEntity;
+import az.devolopia.SpringJava20_Mubariz_Bayramov.entity.UserEntity;
 import az.devolopia.SpringJava20_Mubariz_Bayramov.exception.MyException;
 import az.devolopia.SpringJava20_Mubariz_Bayramov.repository.BookRepository;
 import az.devolopia.SpringJava20_Mubariz_Bayramov.request.BookAddRequest;
+import az.devolopia.SpringJava20_Mubariz_Bayramov.request.BookFilterRequest;
+import az.devolopia.SpringJava20_Mubariz_Bayramov.request.BookFilterRequestForCustomer;
 import az.devolopia.SpringJava20_Mubariz_Bayramov.request.BookUpdateRequest;
 import az.devolopia.SpringJava20_Mubariz_Bayramov.response.BookAddResponse;
 import az.devolopia.SpringJava20_Mubariz_Bayramov.response.BookListResponse;
 import az.devolopia.SpringJava20_Mubariz_Bayramov.response.BookSingleResponse;
 import az.devolopia.SpringJava20_Mubariz_Bayramov.util.MyFileReader;
-
-
 
 @Service
 public class BookService {
@@ -36,9 +39,19 @@ public class BookService {
 	@Autowired
 	private ModelMapper mapper;
 
+	@Autowired
+	private UserService userService;
+
+	@Autowired
+	private MyConfig myConfig;
+
 	public BookAddResponse add(BookAddRequest req) {
 		BookEntity en = new BookEntity();
 		mapper.map(req, en);
+		String username = userService.findUsername();
+		UserEntity operator = userService.findByUsername(username);
+		Integer sellerCode = operator.getUserId();
+		en.setSellerCode(sellerCode);
 		repository.save(en);
 		BookAddResponse resp = new BookAddResponse();
 		resp.setId(en.getId());
@@ -47,8 +60,14 @@ public class BookService {
 	}
 
 	public BookListResponse findAllSearch(String q) {
+		q = q.toLowerCase();
+
+		String username = userService.findUsername();
+		UserEntity operator = userService.findByUsername(username);
+		Integer operatorSellerCode = operator.getUserId();
+
 		BookListResponse s = new BookListResponse();
-		List<BookEntity> filtered = repository.findAllByNameContaining(q);
+		List<BookEntity> filtered = repository.findMyBooksSearch(q, operatorSellerCode);
 		List<BookSingleResponse> list = new ArrayList<BookSingleResponse>();
 
 		for (BookEntity en : filtered) {
@@ -64,7 +83,18 @@ public class BookService {
 		Optional<BookEntity> o = repository.findById(id);
 
 		if (o.isPresent()) {
-			repository.deleteById(id);
+			String username = userService.findUsername();
+			UserEntity operator = userService.findByUsername(username);
+			Integer operatorSellerCode = operator.getUserId();
+
+			BookEntity bookEntity = o.get();
+			Integer bookSellerCode = bookEntity.getSellerCode();
+			if (operatorSellerCode == bookSellerCode) {
+				repository.deleteById(id);
+			} else {
+				throw new MyException("basqasinin kitabini sile bilmezsen", null, "forbidden");
+			}
+
 		} else {
 			throw new MyException("kitab taplmadi id=" + id, null, "id-not-found");
 		}
@@ -84,27 +114,29 @@ public class BookService {
 
 	}
 
-	public void update(BookUpdateRequest u) throws Exception {
+	public void update(BookUpdateRequest u) {
 
 		Integer id = u.getId();
 		Optional<BookEntity> byId = repository.findById(id);
 		if (!byId.isPresent()) {
-
 			String oxunan = fileReader.readFromFile("id-not-found.txt");
 
 			throw new MyException(oxunan, null, "id-not-found");
 		}
-		try {
-			/// gdfgdfg
-		} catch (Exception e) {
-			// TODO: handle exception
-		} finally {
 
-		}
 		BookEntity en = byId.get();
 		mapper.map(u, en);
 
-		repository.save(en);
+		String username = userService.findUsername();
+		UserEntity operator = userService.findByUsername(username);
+		Integer operatorSellerCode = operator.getUserId();
+
+		Integer sellerCode = en.getSellerCode();
+		if (sellerCode == operatorSellerCode) {
+			repository.save(en);
+		} else {
+			throw new MyException("basqasinin kitabini redakte ede bilmezsen", null, "forbidden");
+		}
 
 	}
 
@@ -151,8 +183,67 @@ public class BookService {
 		return s;
 	}
 
-	public Book addBook(Book book) {
-		
-		return null;
+	public BookListResponse findAllSearchFilter(BookFilterRequest r) {
+
+		String username = userService.findUsername();
+		UserEntity operator = userService.findByUsername(username);
+		Integer operatorSellerCode = operator.getUserId();
+
+		BookListResponse s = new BookListResponse();
+		Long count = repository.findMyBooksSearchFilterCheck(operatorSellerCode, r.getName(), r.getId(), r.getPrice(),
+				r.getPageCount(), r.getPublishDate());
+
+		Integer rowCountLimit = myConfig.getRowCountLimit();
+		System.out.println(rowCountLimit);
+		if (count > rowCountLimit) {
+			throw new MyException("axtraisi deiqlqesdirin, tapilan melumat sayisi = " + count + ", maksimum "
+					+ rowCountLimit + " ola biler", null, "data-too-long");
+		}
+
+		List<BookEntity> filtered = repository.findMyBooksSearchFilter(operatorSellerCode, r.getName(), r.getId(),
+				r.getPrice(), r.getPageCount(), r.getPublishDate());
+
+		List<BookSingleResponse> list = new ArrayList<BookSingleResponse>();
+
+		for (BookEntity en : filtered) {
+			BookSingleResponse se = new BookSingleResponse();
+			mapper.map(en, se);
+			list.add(se);
+		}
+		s.setBooks(list);
+		return s;
 	}
+
+	public BookListResponse findAllSearchFilterForCustomer(BookFilterRequestForCustomer r) {
+
+		Integer categoryId = r.getCategoryId();
+		String category = "";
+		if (categoryId != 0) {
+			category = String.valueOf(categoryId);
+		}
+
+		BookListResponse s = new BookListResponse();
+
+		Integer length = r.getLength();
+		if (length > myConfig.getRowCountLimit()) {
+			throw new MyException("melumat limiti asildi", null, "data-too-long");
+		}
+		List<BookEntity> entities = repository.searchFilterForCustomer(r.getName(), category, r.getBegin(),
+				r.getLength());
+
+		Long totalSize = repository.searchFilterCountForCustomer(r.getName(), category);
+
+		List<BookSingleResponse> list = new ArrayList<BookSingleResponse>();
+
+		for (BookEntity en : entities) {
+			BookSingleResponse se = new BookSingleResponse();
+			mapper.map(en, se);
+			list.add(se);
+		}
+		s.setBooks(list);
+		s.setTotalSize(totalSize);
+
+		return s;
+	}
+
 }
